// healthcare-frontend/src/types/tenant.ts
export interface Tenant {

    id?: string;

    name: string;

    slug: string;

    logo?: string;

    website?: string;

    hospital_type?: string;

    city?: string;

    state?: string;

    country?: string;

    about?: string;

    primary_color?: string;

    secondary_color?: string;
}